# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.21.0

package db

import (
	"database/sql"
	"time"
)

type Batch struct {
	ID               int64        `json:"id"`
	Name             string       `json:"name"`
	ActivationStatus string       `json:"activation_status"`
	CustomerID       int64        `json:"customer_id"`
	NoOfDevices      int32        `json:"no_of_devices"`
	DeliveryDate     sql.NullTime `json:"delivery_date"`
	WarrantyEnd      sql.NullTime `json:"warranty_end"`
}

type Bundle struct {
	ID          int64   `json:"id"`
	Mrc         float64 `json:"mrc"`
	Description string  `json:"description"`
}

type BundlesCustomer struct {
	BundlesID   int64 `json:"bundles_id"`
	CustomersID int64 `json:"customers_id"`
}

type Customer struct {
	ID       int64   `json:"id"`
	Customer string  `json:"customer"`
	Paid     float64 `json:"paid"`
	Due      float64 `json:"due"`
}

type Order struct {
	ID        int64           `json:"id"`
	StartDate time.Time       `json:"start_date"`
	EndDate   time.Time       `json:"end_date"`
	BatchID   int64           `json:"batch_id"`
	BundleID  int64           `json:"bundle_id"`
	Nrc       sql.NullFloat64 `json:"nrc"`
}

type PaymentLog struct {
	ID               int64        `json:"id"`
	Payment          float64      `json:"payment"`
	DueDate          time.Time    `json:"due_date"`
	ConfirmationDate sql.NullTime `json:"confirmation_date"`
	OrderID          int64        `json:"order_id"`
	Confirmed        bool         `json:"confirmed"`
	CustomerID       int64        `json:"customer_id"`
}