# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Ticketing support microservice",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Cypodsolutions",
            "url": "http://www.cypod.com/",
            "email": "naders@cypodsolutions.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/batches": {
            "get": {
                "description": "List all batches with optional filter \"customer_name\", pagination params are required",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "List Batches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter: customer name",
                        "name": "customer_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Batch"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new Batch specifying its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Create new Batch",
                "parameters": [
                    {
                        "description": "Create Batch body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Batch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/batches/{batch_id}": {
            "put": {
                "description": "Update Batch by a Batch ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Update Batch",
                "parameters": [
                    {
                        "description": "Update Batch body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateBatchRequestJSON"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Batch ID for update",
                        "name": "batch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Batch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete batch by a batch ID",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Delete Batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "batch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "true"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.createBatchRequest": {
            "type": "object",
            "required": [
                "activation_status",
                "customer_name",
                "name",
                "no_of_devices"
            ],
            "properties": {
                "activation_status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "suspended",
                        "canceled"
                    ]
                },
                "customer_name": {
                    "type": "string",
                    "minLength": 1
                },
                "delivery_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "no_of_devices": {
                    "type": "integer",
                    "minimum": 1
                },
                "warranty_end": {
                    "type": "string"
                }
            }
        },
...
        "api.updateBatchRequestJSON": {
            "type": "object",
            "properties": {
                "activation_status": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "delivery_date": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "no_of_devices": {
                    "type": "integer"
                },
                "warranty_end": {
                    "$ref": "#/definitions/sql.NullTime"
                }
            }
        },
...
        "db.Batch": {
            "type": "object",
            "properties": {
                "activation_status": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "delivery_date": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "no_of_devices": {
                    "type": "integer"
                },
                "warranty_end": {
                    "$ref": "#/definitions/sql.NullTime"
                }
            }
        },
...
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}